Rename gnark_backend_glue to gnark_backend_wrapper
