+++ src/gnark_backend_wrapper/errors.rs
