+++ src/lib.rs
@@ -3,4 +3,4 @@ mod acvm;
 pub mod acvm_interop;
 pub use acvm_interop::Gnark;
 
-mod gnark_backend_glue;
+mod gnark_backend_wrapper;
